Move ConvertingSerializer implementation into main module of hyperon-atom
