+++ hyperon-atom/src/serial.rs
@@ -36,39 +36,6 @@ pub enum Error {
     NotSupported,
 }
 
-/// Converts serializable grounded atom into native Rust type `T`.
-/// Returning error from this serializer is not expected as usually it should
-/// just write data into a memory buffer for conversion. [Default] is required
-/// to instantiate it in [ConvertingSerializer::convert] method.
-pub trait ConvertingSerializer<T>: Serializer + Default {
-    fn into_type(self) -> Option<T>;
-
-    /// Converts atom into Rust value using `Self::default()` instance.
-    /// First it checks whether the grounded atom is already an instance of `T`
-    /// and clones value if it the case. Otherwise it uses `Self` to convert
-    /// into `T` via serialization.
-    fn convert(atom: &super::Atom) -> Option<T>
-        where T: 'static + Clone
-    {
-        std::convert::TryInto::<&dyn super::GroundedAtom>::try_into(atom)
-            .ok()
-            .and_then(|gnd| {
-                gnd.as_any_ref()
-                    // TODO: it is not clear whether this step really improves performance.
-                    // On the other hand it requires `T` to be static and implement `Clone`.
-                    // It is better to do a performance test and check if first step should be
-                    // removed.
-                    .downcast_ref::<T>()
-                    .cloned()
-                    .or_else(|| {
-                        let mut serializer = Self::default();
-                        gnd.serialize(&mut serializer).ok()
-                            .and_then(|()| serializer.into_type())
-                    })
-            })
-    }
-}
-
 /// Serialization result type
 pub type Result = std::result::Result<(), Error>;
 
@@ -117,24 +84,6 @@ mod tests {
 
     use std::fmt::Display;
 
-    #[derive(Default)]
-    struct I64Serializer {
-        value: Option<i64>,
-    }
-
-    impl Serializer for I64Serializer {
-        fn serialize_i64(&mut self, v: i64) -> super::Result {
-            self.value = Some(v);
-            Ok(())
-        }
-    }
-
-    impl ConvertingSerializer<i64> for I64Serializer {
-        fn into_type(self) -> Option<i64> {
-            self.value
-        }
-    }
-
     #[derive(PartialEq, Debug, Clone)]
     struct I64Gnd(i64);
 
@@ -152,12 +101,6 @@ mod tests {
         fn serialize(&self, serializer: &mut dyn serial::Serializer) -> serial::Result {
             serializer.serialize_i64(self.0)
         }
-    }
-
-    #[test]
-    fn convert_return_none_on_incorrect_grounded_atom() {
-        assert_eq!(I64Serializer::convert(&Atom::gnd(I64Gnd(42))), Some(42));
-        assert_eq!(I64Serializer::convert(&Atom::value("42")), None);
     }
 
     #[derive(PartialEq, Debug, Clone)]
